requirements.txt


python-dotenv
pandas
numpy
requests
matplolib.pyplot
seaborn
plotly.express
plotly.graph_objects
datetime